hordelib/nodes/comfy_controlnet_preprocessors/requirements.txt
opencv-contrib-python
opencv-python
timm==0.6.12
torchvision
scipy
matplotlib
scikit-image
prettytable==3.6.0
svglib
reportlab
addict
yapf
mediapipe==0.9.1.0